rfactor: Bring subGraph back to Task from GraphAsTask (TaskWithSubgraph)
- execute has ownership, so no automatic run, basic task runner ignores it, json export ignores it, etc.
